+++ packages/core/src/GameEngine/index.ts
@@ -1,0 +1,1 @@
+export { GameEngineConfig, defaultConfig, GameEngine } from "./GameEngine"
