+++ packages/core/src/index.ts
@@ -1,19 +1,9 @@
-export interface Vector2 {
-    x: number,
-    y: number
-}
+export {
+    Vector2
+} from "./types"
 
-export interface GameEngineConfig {
-    referenceResoulution: Vector2
-}
-
-export const defaultConfig: GameEngineConfig = {
-    referenceResoulution: { x: 4096, y: 4096 }
-}
-
-export class GameEngine {
-    constructor(config: GameEngineConfig) {
-        console.log("Game Engine Instantiated")
-        console.log("Reference Resolution: ", config.referenceResoulution)
-    }
-}
+export {
+    GameEngineConfig,
+    defaultConfig,
+    GameEngine
+} from "./GameEngine"
