+++ packages/core/src/types/Vector2.ts
@@ -1,0 +1,4 @@
+export interface Vector2 {
+    x: number,
+    y: number
+}
