+++ packages/core/src/types/index.ts
@@ -1,0 +1,1 @@
+export { Vector2 } from "./Vector2"
