+++ turbo.json
@@ -10,6 +10,9 @@
             ]
         },
         "test": {
+            "dependsOn": [
+                "^build"
+            ],
             "outputs": [
                 "coverage/**"
             ]
